packages/openseneca/openseneca-0.1.1-py3-none-any.whl/tests/test_gpt3.py
import pytest
from unittest.mock import Mock, patch
from openseneca.models.gpt3 import GPT35ChatModel

# TODO: need more details to test the GPT35ChatModel class

def test_gpt35_chat_model_initialization():
    provider_mock = Mock()
    model = GPT35ChatModel(provider=provider_mock)
    assert model.provider == provider_mock

def test_gpt35_chat_model_request():
    provider_mock = Mock()
    provider_mock.request.return_value = "Test response"
    model = GPT35ChatModel(provider=provider_mock)

    response = model.request("Test prompt")
    assert response == "Test response"
    provider_mock.request.assert_called_once()

@patch('openseneca.models.gpt3.GPT35ChatModel.request')
def test_gpt35_chat_model_request_with_instructions(mock_request):
    mock_request.return_value = "Test response"
    model = GPT35ChatModel(provider=Mock())
    response = model.request("Test prompt", instructions="Test instructions")
    assert response == "Test response"
    mock_request.assert_called_once()

@patch('openseneca.models.gpt3.GPT35ChatModel.request')
def test_gpt35_chat_model_request_with_past_messages(mock_request):
    mock_request.return_value = "Test response"
    model = GPT35ChatModel(provider=Mock())
    response = model.request("Test prompt", past_messages=["Past message"])
    assert response == "Test response"
    mock_request.assert_called_once()

@patch('openseneca.models.gpt3.GPT35ChatModel.request')
def test_gpt35_chat_model_request_with_settings(mock_request):
    mock_request.return_value = "Test response"
    model = GPT35ChatModel(provider=Mock())
    response = model.request("Test prompt", settings={"setting": "value"})
    assert response == "Test response"
    mock_request.assert_called_once()